Julian dates be in double precision
`ConfigureCompilerFlags.cmake` and `TAGS.cmake` are now copied into
LMDZE, to avoid dependency on the environment.

Julian dates must be in double precision, to get time step precision.

Add optional attribute to argument sec of procedure ju2ymds. We do
not need sec in procedure dynredem0.

In procedure ju2ymds, by construction, sec cannot be > `un_jour`.

Remove useless intermediary variables in procedure ymds2ju.

+++ IOIPSL/Histcom/histend.f90
@@ -33,7 +33,7 @@ contains
     INTEGER dims(4), dim_cnt
     INTEGER year, month, day, hours, minutes
     REAL sec
-    REAL rtime0
+    double precision rtime0
     CHARACTER(len=20) tname, tunit
     CHARACTER(len=42) str30
     CHARACTER(len=80) ttitle
